chore: Imporve styling of AddRow dialogs and autocomplete lists.
