--- src/components/StrategraphicColumn/StratigraphicColumnSelect/StratigraphicColumnSelect.styled.tsx
@@ -1,9 +1,0 @@
-import styled from 'styled-components';
-import { spacings } from '../../../tokens/spacings';
-
-export const AutocompleteList = styled.div`
-  display: flex;
-  flex-direction: column;
-
-  row-gap: ${spacings.MEDIUM};
-`;
