+++ src/features/ModelView/ModelMetadataView/ModelMetadataView.styled.tsx
@@ -1,4 +1,3 @@
-import { Dialog } from '@equinor/eds-core-react';
 import styled from 'styled-components';
 import { spacings } from '../../../tokens/spacings';
 
@@ -25,9 +24,4 @@ export const MetadataInfo = styled.div`
   row-gap: ${spacings.MEDIUM};
 
   max-width: 60%;
-`;
-
-export const Actions = styled(Dialog.Actions)`
-  display: flex;
-  column-gap: ${spacings.SMALL};
 `;
